Add ServerName component
